Restructuring files to better reflect functions and tests grouping
