+++ redux_reader.go
@@ -1,17 +1,17 @@
 package kvas
 
 import (
-	"github.com/boggydigital/wits"
 	"golang.org/x/exp/maps"
 	"golang.org/x/exp/slices"
-	"io"
-	"sort"
-	"strings"
 )
+
+func ReduxReader(dir string, assets ...string) (ReadableRedux, error) {
+	return connectRedux(dir, assets...)
+}
 
 func (rdx *Redux) MustHave(assets ...string) error {
 	for _, asset := range assets {
-		if !rdx.Has(asset) {
+		if !rdx.HasAsset(asset) {
 			return UnknownReduxAsset(asset)
 		}
 	}
@@ -22,7 +22,7 @@ func (rdx *Redux) Keys(asset string) []string {
 	return maps.Keys(rdx.assetKeyValues[asset])
 }
 
-func (rdx *Redux) Has(asset string) bool {
+func (rdx *Redux) HasAsset(asset string) bool {
 	_, ok := rdx.assetKeyValues[asset]
 	return ok
 }
@@ -49,7 +49,7 @@ func (rdx *Redux) HasValue(asset, key, val string) bool {
 }
 
 func (rdx *Redux) GetAllValues(asset, key string) ([]string, bool) {
-	if !rdx.Has(asset) {
+	if !rdx.HasAsset(asset) {
 		return nil, false
 	}
 	if rdx.assetKeyValues[asset] == nil {
@@ -65,160 +65,4 @@ func (rdx *Redux) GetFirstVal(asset, key string) (string, bool) {
 		return values[0], true
 	}
 	return "", false
-}
-
-func (rdx *Redux) assetModTime(asset string) (int64, error) {
-	return rdx.kvr.CurrentModTime(asset)
-}
-
-func (rdx *Redux) ModTime() (int64, error) {
-	rdx.mtx.Lock()
-	defer rdx.mtx.Unlock()
-
-	var mt int64 = 0
-	for asset := range rdx.assetKeyValues {
-		if amt, err := rdx.assetModTime(asset); err != nil {
-			return -1, err
-		} else {
-			if mt < amt {
-				mt = amt
-			}
-		}
-	}
-	return mt, nil
-}
-
-func (rdx *Redux) Refresh() error {
-	if err := rdx.kvr.IndexRefresh(); err != nil {
-		return err
-	}
-
-	modTime, err := rdx.ModTime()
-	if err != nil {
-		return err
-	}
-
-	if rdx.modTime < modTime {
-		if rdx, err = connectRedux(rdx.dir, maps.Keys(rdx.assetKeyValues)...); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (rdx *Redux) MatchAsset(asset string, terms []string, scope []string, options ...MatchOption) []string {
-	if scope == nil {
-		scope = rdx.Keys(asset)
-	}
-
-	matches := make(map[string]interface{})
-	for _, term := range terms {
-		if !slices.Contains(options, CaseSensitive) {
-			term = strings.ToLower(term)
-		}
-		for _, key := range scope {
-			if values, ok := rdx.GetAllValues(asset, key); !ok {
-				continue
-			} else if anyValueMatchesTerm(term, values, options...) {
-				matches[key] = nil
-			}
-		}
-	}
-
-	return maps.Keys(matches)
-}
-
-type MatchOption int
-
-const (
-	CaseSensitive = iota
-	FullMatchOnly
-)
-
-func (rdx *Redux) Match(query map[string][]string, options ...MatchOption) []string {
-	var matches []string
-	for asset, terms := range query {
-		if !rdx.Has(asset) {
-			continue
-		}
-		matches = rdx.MatchAsset(asset, terms, matches, options...)
-	}
-	return matches
-}
-
-func anyValueMatchesTerm(term string, values []string, options ...MatchOption) bool {
-	anyCase := true
-	contains := true
-
-	if options != nil {
-		anyCase = !slices.Contains(options, CaseSensitive)
-		contains = !slices.Contains(options, FullMatchOnly)
-	}
-
-	for _, val := range values {
-		if anyCase {
-			val = strings.ToLower(val)
-		}
-		if contains {
-			if strings.Contains(val, term) {
-				return true
-			}
-		} else {
-			if val == term {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func (rdx *Redux) Sort(ids []string, desc bool, sortBy ...string) ([]string, error) {
-	if err := rdx.MustHave(sortBy...); err != nil {
-		return nil, err
-	}
-
-	sis := &sortableIdSet{
-		properties: sortBy,
-		ipv:        make([]idValues, 0, len(ids)),
-	}
-
-	for _, id := range ids {
-		iv := idValues{id: id}
-		for _, p := range sortBy {
-			v, _ := rdx.GetFirstVal(p, id)
-			iv.values = append(iv.values, v)
-		}
-		sis.ipv = append(sis.ipv, iv)
-	}
-
-	var sortInterface sort.Interface = sis
-	if desc {
-		sortInterface = sort.Reverse(sortInterface)
-	}
-
-	sort.Sort(sortInterface)
-
-	sorted := make([]string, 0, len(sis.ipv))
-	for _, iv := range sis.ipv {
-		sorted = append(sorted, iv.id)
-	}
-
-	return sorted, nil
-}
-
-func (rdx *Redux) Export(w io.Writer, ids ...string) error {
-
-	skv := make(wits.SectionKeyValues)
-
-	for _, id := range ids {
-		skv[id] = make(wits.KeyValues)
-		for p := range rdx.assetKeyValues {
-			if vals, ok := rdx.GetAllValues(p, id); ok {
-				skv[id][p] = vals
-			}
-		}
-	}
-
-	return skv.Write(w)
 }
